Summary
• Stabilizes runtime by fixing event store acquisition and trigger
wiring.
• Corrects RabbitMQ routing key construction to ensure proper message
routing.
•  Fixes configuration loading/deserialization for node properties.
• Hardens event parameter handling: corrects indexed parameter
extraction and reduces decoder fragility.
• Migrates numeric event parameter types from Integer to BigInteger
(domain, persistence JPA/Mongo, and tests) to prevent overflow and
precision issues.

Key changes
• Event store: obtain the correct instance in both node initialization
and broadcaster trigger paths, preventing NPEs and misrouted operations.
• Messaging: build RabbitMQ routing keys correctly in
RabbitMqBroadcasterProducer.
• Config: fix NodeProperties deserialization so configuration loads
reliably at startup.
• Parameters/decoding: fix indexed parameter extraction and streamline
DefaultContractEventParameterDecoder to avoid incorrect decoding.
• Types: switch Int/Uint parameter representations and persistence
entities/documents to BigInteger; update unit tests accordingly.

+++ core/src/test/java/io/naryo/application/node/routing/EventRoutingServiceTest.java
@@ -1,5 +1,6 @@
 package io.naryo.application.node.routing;
 
+import java.math.BigInteger;
 import java.util.*;
 
 import io.naryo.application.configuration.revision.Revision;
@@ -117,7 +118,7 @@ class EventRoutingServiceTest {
         when(ce.getEventType()).thenReturn(EventType.CONTRACT);
         when(ce.getName()).thenReturn(new EventName("Foo"));
         when(ce.getStatus()).thenReturn(ContractEventStatus.CONFIRMED);
-        IntParameter param = new IntParameter(false, 0, 42);
+        IntParameter param = new IntParameter(false, 0, BigInteger.valueOf(42));
         when(ce.getParameters()).thenReturn(Set.of(param));
 
         Revision<Broadcaster> broadcasterRevision =
